imagex--fit-to-size: Handle case of nil maximum magnification
Because the imagex--fit-to-size function takes an optional parameter
for the maximum magnification and (min ...) has problems if one of the
arguments is nil, the magnification calculation has been changed to
deal with nil parameters.
Handle errors in case the file has been renamed or moved

Conflicts:
	image+.el

not enough consideration but will be fixed in future commit.